Mount host's root filesystem into CSI plugin container
If we want to allow the cluster administrator to specify arbitrary
btrfs subvolume locations (via the
CSIDriver.spec.parameters.subvolumeRoot parameter), then we need to
have all possible host directories available in the container.
At the same time, this means we don't need to have btrfs tooling in
the container, since we can just chroot to the host.